Add a section on error handling and recovery.
I haven't checked the OS error reporting interfaces to see if this is
implementable. The security upgrades are likely to be totally under the control
of the OS rather than the browser, but I think this wording will allow the
browsers to just follow their OS.

Fixes #25 to the extent we can do so.