updated charter to match template updates
These are changes to make it clearer that the formal parts of the process are for when the usual informal process fails -- and that is considered rare.  That's mostly by saying that and breaking up long paragraphs so the only if bad things happen parts are separated out.

There are a few substantive differences between this and the template, but I didn't put those in.  (minor things, like there's a minimum number of people required to set off those formal processes in the template, not 1 as here. 

It's find to drop this if it's too late.  It's mostly just for clarity.