Throw NotFoundError when trying to look up a non-existent attribute.
The common case seems to be navigating a tree with certain attributes expected
to be there, and null checks make that more complicated. It's still possible to
handle optional attributes by catching the exception.

We have less experience with the multi-attribute getters, but I'm doing them
consistently for now.

Fixes #124.