Remove the device info and device class fields from BluetoothDevice.
These fields are in the API because it started life as
chrome.bluetooth.Device, but they don't exist in either the Android or
Core Bluetooth APIs. Developers can still navigate the GATT tree to look
them up; they just won't be available at the top level anymore.

I asked for opinions on doing this at
https://lists.w3.org/Archives/Public/public-web-bluetooth/2016Apr/0017.html
and the limited response was positive.

This change also makes the "get the BluetoothDevice representing a
Bluetooth device" algorithm synchronous, which will help implement #163.