Define Bluetooth tree and cache
Explicitly define Bluetooth tree and Bluetooth cache so that references
to it are linked properly. The paragraph describing the Bluetooth tree
was reworded so that Bluetooth tree was at the start of the paragraph so
that links will display the entire paragraph instead of just the end
where the definition was originally.