Use Markdown for Sections 6-9 (#473)
* Use Markdown for Sections 6-9

This change updates the spec source file to use more Markdown syntax in
order to improve readability. This commit updates Sections 6 to 9 of the
spec. Most of Section 9 still maintains its list in HTML, since it uses
specific numbering and lists with letters.

The following changes were performed:

* Heading tags were replaced with their Markdown equivalent, `#`.
* `<p>` tags were removed.
* `<ol>`, `<ul>`, and `<li>` tags were replaced with their Markdown
    equivalent, `*` and `1.`.
* `<table>` tags had the `data` class removed to fix their rendering
    such that they don't overflow past their containers.

* Define Bluetooth tree and cache

Explicitly define Bluetooth tree and Bluetooth cache so that references
to it are linked properly. The paragraph describing the Bluetooth tree
was reworded so that Bluetooth tree was at the start of the paragraph so
that links will display the entire paragraph instead of just the end
where the definition was originally.