Fix visibility change and loss of full activity events
This change fixes the steps for handling visibility change and loss of
full activity to find the global object for the Document which
experienced the event and from there identify the Bluetooth object for
which we need to take action.

This introduces the concept of an "associated Bluetooth", similar to an
"associated Document" and adds a "Bluetooth task source" which is now
used whenever queuing a task.

Fixed #593.