Rewrite index.bs to always queue a task before resolving a Promise
This fixes the algorithms identified by #641 as resolving or rejecting a
Promise while running in parallel without first queuing a task.

There are a lot of other issues with the algorithms in this spec and
while I've cleaned up the text immediately around the affected
algorithms much has been left as-is to avoid rewriting the world.

Fixed #641.